Primordial soup
The first idea for the app. You can select the key, the number of strings up to 6 strings and choose between a major and minor chord.
You can also select the tuning just by typing the letters for the root notes needed. If the number of notes is different from the number of strings it will result in an error and the fretboard will not be displayed.

It's ugly and the code looks a mess but hey! It works.

- For now sharp or flat root notes are not permitted.